requirements.txt
django==3.1.7
djangorestframework==3.12.4
IPython
Pillow==8.2.0
numpy==1.20.2
